requirements.txt
python-telegram-bot==13.3
instagrapi
requests==2.31.0
